Compute the URLs of assets from the request
This allows us to deploy to any URL without having to specify the base
address in the server.xml

Note that this changes the semantics of the constructors of subclasses
of RepositoryObject. Previously RepositoryObject(Map<String, Object>
properties) would create a copy of the properties map but now it does
not. This makes performing business logic on a list of objects much
easier. (Previously, when iterating through an AttachmentList, any
changes made were not reflected in the original list.)

Some parts of the code have been updated to manually make a copy where
appropriate.

Conflicts:
	server/src/main/java/com/ibm/ws/lars/rest/RepositoryRESTResource.java