Add asset reviews function
Reviews were an old feature which we used to show on the website but
were later removed. We anticipate eventually supporting reviews again
when we allow third-parties to add stuff to the repo.

For now, we want a stub implementation of the assetreviews resource in
LARS.

The website makes calls to /ma/v1/assets/{id}/assetreviews (e.g.
https://asset-websphere.ibm.com/ma/v1/assets/5502f073d759edd245000067/assetreviews?apiKey=75621234192)

This allows LARS to handle this request, always returning an empty JSON
array.