Add a front page to LARS displaying the server status
Add a new simple servlet which displays a welcome page with the number
of assets in the server. This serves as a simple IVT and also allows
monitoring tools to work in the default configuration where they attempt
to retrieve the application root.

Adding the new servlet required moving the root of the JAX-RS
application from '/' to '/ma/v1' which resulted in some additional code
changes.

As a nice side-effect, I also made the BaseURL configuration a little
more friendly so that it will work whether or not it ends with a slash.

Updated the method names on Configuration to better clarify what they
actually do.

Changed AssetServiceLayer from `Singleton` to `ApplicationScoped`.
Singleton is not a CDI scope, so presumably AssetServiceLayer was being
injected as a dependant bean which is not what we want, especially if
anyone actually tries to Serialize the FrontPageServlet.