[BUG] Handle nulls in JSON arrays
This fix handles the bug where a null is in a JSON array that is being
deserialized.  It also tidies up the code within the
DataModelSerializer.processJsonArray to remove some of the nesting.

Tests are added for deserializing of nulls and nulls in lists as well as
clarifying the support for JSON true/false.