AddThenDeleteStrategy should still update state of unchanged asset
When uploading an asset with AddThenDeleteStrategy (or any subclass), it
checks to see whether the asset already exists and if so, whether the
new asset has any differences.

If there are no differences between the old and new asset, it will not
do the upload since it would result in no change. However, in this case,
it should still set the existing asset to the desired state passed in in
the constructor.