Fix to properly delete attachment content
Make sure that attachment content is actually removed from the mongo
database when the attachment is deleted.

Add check to ensure attachment content is deleted

Change Persistor.deleteAttachmentContent to take the gridFsId rather
than the attachment id. This matches the signature for
retrieveAttachmentMetadata and the difference is probably why the bug
wasn't originally spotted.

PersistenceBean.deleteAttachmentContent now actually does what it says
it does.

AssetServiceLayer change to match the new semantics of
deleteAttachmentContent.

RepositoryContext change to check for orphaned attachment content when
cleaning repository between tests.