Fix to correctly delete attachments when asset is deleted
Add test that deleting asset deletes attachments

Also add a check to ensure there are no attachments remaining in the
database after clearing the repository at the end of each test.

Note that previously, the code was trying to retrieve the attachments to
delete by calling asset.getAttachments(). However, this doesn't work
because the previous call to retrieveAssets() does not populate the
asset object with its attachments. This is by design because we don't
always want to return the attachments associated with an asset.