Reduce memory usage when returning large result sets
This change allows results to be streamed from the database into a
response to the client when retrieving multiple assets. It avoids
needing to load the whole result set into memory before returning it.

To do this we introduce the AssetCursor, which wraps a database result
cursor and allows the Rest layer to iterate through the results without
needing to know the details of the persistence backend.

To allow changes to be made to the information from the database before
it is consumed by the rest layer we introduce the AssetOperation, which
encapsulates some processing which should be done to the asset before it
is returned (e.g. transforming between data types or removing
implementation-specific fields)