Add test for the behaviour of an empty filter
If a filter is specified with an empty value, it is treated as
though filtering for assets which an empty value for the property.
Add a test for this behaviour, as it wasn't previously tested.
N.B. This behavious is different to the behaviour of the legacy
Massive repository server, but no clients rely on the old behaviour
so it should be safe to change it.