Create SingleFileClient and SingleFileRepositoryConnection
SingleFileClient holds the parsed JSON, Assets are then parsed as
required. This is a good fit for the contract of RepositoryClient which
was originally designed for retrieval of assets from a REST service.

Extend the DataModelSerializer to allow serializing to and from
JsonObjects as well as streams.

Extend the RepositoryClientTest to run against the SingleFileClient.

Add explicit test that IDs are maintained as SingleFileClient generates
IDs based on the position in the file.